ChartZoom on own folder and files separate | implement animation
